# docs/openapi/analysis-datashield-adaptor/openapi.json
{
  "openapi" : "3.1.0",
  "paths" : {
    "/ws/datasource/{datasource}/table/{table}" : {
      "get" : {
        "parameters" : [ {
          "name" : "datasource",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "table",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Select Table",
        "tags" : [ "WS Resource" ]
      }
    },
    "/ws/system/subject-profile/{profile}" : {
      "get" : {
        "parameters" : [ {
          "name" : "profile",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Accept-Encoding",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Content-Length",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Host",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "User-Agent",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        },
        "summary" : "Get Subject Profile",
        "tags" : [ "WS Resource" ]
      }
    }
  },
  "info" : {
    "title" : "DataSHIELD Invoker API (production)",
    "version" : "1.0.1",
    "description" : "Just an example service",
    "termsOfService" : "Your terms here",
    "contact" : {
      "name" : "Example API Support",
      "email" : "techsupport@example.com",
      "url" : "http://exampleurl.com/contact"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers" : [ {
    "url" : "https://localhost:8543",
    "description" : "Auto generated value"
  }, {
    "url" : "https://0.0.0.0:8543",
    "description" : "Auto generated value"
  } ]
}